BUG#: 6098 TITLE: Use Uint32 instead of size_t in Buffer class DESCRIPTION: Use Uint32 in the Buffer interface to remove the need for explicit type conversions from size_t to Uint32.

+++ src/Clients/repupgrade/RepositoryUpgrade.cpp
@@ -2036,11 +2036,11 @@ void RepositoryUpgrade::_logRequestToFile(
 
             if( contentLength > 0 )
             {
-                contentOffset = (Uint32)data.size() - contentLength;
+                contentOffset = data.size() - contentLength;
             }
             else
             {
-                contentOffset = (Uint32)data.size();
+                contentOffset = data.size();
             }
 
             //
